requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# EDA Agent — Python Dependencies
# Install with:  pip install -r requirements.txt
# Python 3.10+ required
# ──────────────────────────────────────────────────────────────────────────────

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Streamlit web interface
streamlit>=1.35.0

# Generative AI (Gemini)
google-genai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Interactive HTML dashboard
plotly>=5.18.0
jinja2>=3.1.0

# (Optional) Faster CSV parsing — significant speed boost on large files
# Uncomment if desired:
# pyarrow>=14.0.0

# ── Development & Testing ─────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0     # test coverage raporu için

